feat: Added view helper
